Automate maintenance of generic_COBALT diagnostics
To simplify maintenance of registering diagnostics (in the
generic_COBALT_register_diag subroutine), I've moved the info for the
COBALT diagnostics to a .yaml file.  I also added a small python script
to generate the necessary source code from that yaml file and dump
it to a separate file, which can be regenerated as needed when new
diagnostics are added or diagnostic-registering syntax is changed.