Replace most calls to g_tracer_add_param
Add a call to read the name and loaction of the COBALT param files
from the namelist to generic_COBALT

Replace calls to g_tracer_add_param with get_param but do not change deaults.
The descriptions are just the parameter name, and the units are included where
they were documenemted in the code.