Add support for a COBALT_param_doc file (#59)
* Add cobalt_param_doc.F90

Using the file parser from MOM6, add the option to read parameters
from COBALT_input and COBALT_override files and write pramerters to
COBALT_param_doc files. The COBALT_input files must be specific in
the input.nml

* Replace most calls to g_tracer_add_param

Add a call to read the name and loaction of the COBALT param files
from the namelist to generic_COBALT

Replace calls to g_tracer_add_param with get_param but do not change deaults.
The descriptions are just the parameter name, and the units are included where
they were documenemted in the code.

* Add support for a COBALT_param_doc file

Using the file parser from MOM6, add the option to read parameters
from COBALT_input and COBALT_override files and write pramerters to
COBALT_param_doc files. The COBALT_input files must be specific in
the input.nml

* Replace most calls to g_tracer_add_param

Replace calls to g_tracer_add_param with get_param but do not change deaults.
The descriptions are just the parameter name, and the units are included where
they were documenemted in the code.

* Fix typo

Fix error in line continue statement.

* Add missing description

Add description for RHO_0 and NKML

* Cange parameters with sperd or spery

Remove unit changes from per-day or per-year to per-second from
the default values to do that calculation outside the get param
call. This should not change the solution.

* Use scale argument in get_param

Where default values were rescaled by time (sperd or spery), nutrients
(carbon to nitrogen), concentration (micromoles to moles) or light, the
optional argument scale is now set in get_param so that the values in
the parameter doc files are consistent with the literature.

---------

Co-authored-by: Theresa Morrison <Theresa.Morrison@gaea55.ncrc.gov>
Co-authored-by: Theresa Morrison <Theresa.Morrison@gaea57.ncrc.gov>
Co-authored-by: Theresa Morrison <Theresa.Morrison@ldt-4788475.gfdl.noaa.gov>